requirements.txt
Flask==3.0.2
mysql-connector-python==8.3.0
selenium==4.21.0
webdriver_manager==4.0.1